IO reorganization
Changed '_save_attrs_' to '_io_attrs_' becuase it is used for both saving and loading.
Moved 'comput_things' functions from configuration.py into their own file: compute_funs.py.
Moved 'unpack_state' from configuration.py to backend.py. to avoid circular import statements.
Fixed bug in reading/writing Equilibrium initial guess.
