+++ tests/test_input_output.py
@@ -8,16 +8,6 @@ from desc.configuration import Configuration, Equilibrium
 #from desc.input_output import read_input
 
 
-#class TestIO(unittest.TestCase):
-#    """tests for input/output functions"""
-#
-#    def test_min_input(self):
-#        dirname = os.path.dirname(__file__)
-#        filename = os.path.join(dirname, 'MIN_INPUT')
-#        inputs = read_input(filename)
-#
-#        self.assertEqual(len(inputs), 26)
-
 class TestInputReader(unittest.TestCase):
 
     def setUp(self):
@@ -54,7 +44,6 @@ class TestInputReader(unittest.TestCase):
             'variable incorrect with default argument')
         self.assertFalse(ir.args.version, 'version is not default False')
         self.assertEqual(len(ir.inputs), 28, 'number of inputs does not match '
-
             'number expected in MIN_INPUT')
         # test equality of arguments
 
@@ -80,9 +69,12 @@ class TestInputReader(unittest.TestCase):
     def test_vmec_to_desc_input(self):
         pass
 
+
 class MockObject:
+
     def __init__(self):
-        self._save_attrs_ = ['a', 'b', 'c']
+        self._io_attrs_ = ['a', 'b', 'c']
+
 
 class Testhdf5Writer(unittest.TestCase):
 
@@ -140,20 +132,21 @@ class Testhdf5Writer(unittest.TestCase):
             writer.write_obj(mo)
         writer.close()
         writer = hdf5Writer(self.filename, self.file_mode)
-        for name in mo._save_attrs_:
+        for name in mo._io_attrs_:
             setattr(mo, name, name)
         writer.write_obj(mo)
         groupname = 'initial'
         writer.write_obj(mo, where=writer.sub(groupname))
         writer.close()
         f = h5py.File(self.filename, 'r')
-        for key in mo._save_attrs_:
+        for key in mo._io_attrs_:
             self.assertTrue(key in f.keys())
         self.assertTrue(groupname in f.keys())
         initial = f[groupname]
-        for key in mo._save_attrs_:
+        for key in mo._io_attrs_:
             self.assertTrue(key in initial.keys())
         f.close()
+
 
 class Testhdf5Reader(unittest.TestCase):
 
@@ -218,13 +211,13 @@ class Testhdf5Reader(unittest.TestCase):
         mo = MockObject()
         reader = hdf5Reader(self.filename)
         reader.read_obj(mo)
-        mo._save_attrs_  += '4'
+        mo._io_attrs_  += '4'
         with self.assertWarns(RuntimeWarning):
             reader.read_obj(mo)
-        del mo._save_attrs_[-1]
+        del mo._io_attrs_[-1]
         submo = MockObject()
         reader.read_obj(submo, where=reader.sub(self.subgroup))
-        for key in mo._save_attrs_:
+        for key in mo._io_attrs_:
             self.assertTrue(hasattr(mo, key))
             self.assertTrue(hasattr(submo, key))
